+++ src/count.ts
